refactor(dts): split declared call return gate
